// src/app/features/home/models/detail-data.ts
export interface DetailData {
  act: string;
  ar: string;
  aren: string;
  available_rent_bikes: number;
  available_return_bikes: number;
  infoDate: string;
  infoTime: string;
  latitude: number;
  longitude: number;
  mday: string;
  sarea: string;
  sareaen: string;
  sna: string;
  snaen: string;
  sno: string;
  srcUpdateTime: string;
  total: number;
  updateTime: string;
  label?: string;
}